Add a preliminary concept to instrument xdp progs
In this proof-of-concept, we write a simple XDP program called
udp_prog. The program doesn't do anything except modify the TOS field
in the ipv4 header.

However, instead of attaching this XDP program normally, we instead
wrap it with our own bpf_scope_xdp program. This program first
calculates a packet checksum, then calls the given XDP program. Then,
it recomputes the checksum to see if anything in the packet
changed. If it did, it keeps track of this in a bpf hash map.

Each network flow is identified in the bpf hash map by its IP
5-tuple.